pyproject.toml
[project]
name = "manga-ocr"
authors = [
    {name = "Maciej Budyś", email = "kha-white@mail.com"},
]
description = "OCR for Japanese manga"
readme = "README.md"
requires-python = ">=3.6"
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python :: 3",
]

dynamic = ["version"]
dependencies = [
    "fire",
    "fugashi",
    "jaconv",
    "loguru",
    "numpy",
    "Pillow>=10.0.0",
    "pyperclip",
    "torch>=1.0",
    "transformers>=4.25.0",
    "unidic_lite",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
]


[build-system]
requires = ["setuptools >= 61.0.0", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["manga_ocr"]

[tool.setuptools.dynamic]
version = {attr = "manga_ocr._version.__version__"}

[tool.ruff]
line-length = 120
indent-width = 4

[project.scripts]
manga_ocr = "manga_ocr.__main__:main"